requirements.txt
Django==2.2.6
django-rest-auth==0.9.5
djangorestframework==3.10.3
PyYAML==5.1.2
dynaconf==2.1.1
coreapi==2.3.3
psycopg2==2.8.3
dj-database-url==0.5.0
gunicorn==19.9.0
whitenoise==4.1.4
drf-yasg==1.17.0